pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    # ComplexFuncBench dependencies
    "torch>=2.3.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "scipy>=1.11.0",
    "anthropic>=0.25.0",
    "loguru>=0.7.0",
    "FlagEmbedding>=1.3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
